refactor: remove resemble engine and make pixelmatch the sole comparison engine
