+++ packages/image-comparison-core/src/helpers/options.interfaces.ts
@@ -206,15 +206,6 @@ export interface ClassOptions {
     scaleImagesToSameSize?: boolean;
 
     /**
-     * The image comparison engine to use.
-     * 'pixelmatch' is a beta alternative that uses the YIQ color space and is
-     * more robust against font rendering noise and anti-aliasing differences.
-     * Prefer setting this inside compareOptions.
-     * @default 'resemble'
-     */
-    compareEngine?: 'resemble' | 'pixelmatch';
-
-    /**
      * Options object passed to the underlying image comparison engine.
      */
     compareOptions?: Partial<CompareOptions>;
@@ -491,12 +482,5 @@ export interface CompareOptions {
      * Scale images to the same size before comparing them.
      */
     scaleImagesToSameSize: boolean;
-    /**
-     * The image comparison engine to use.
-     * 'pixelmatch' is a beta alternative that uses the YIQ color space and is
-     * more robust against font rendering noise and anti-aliasing differences.
-     * @default 'resemble'
-     */
-    compareEngine?: 'resemble' | 'pixelmatch';
 }
