+++ packages/image-comparison-core/src/methods/compareReport.interfaces.ts
@@ -1,4 +1,4 @@
-import type { CompareData } from 'src/resemble/compare.interfaces.js'
+import type { CompareData } from 'src/pixelmatch/compare.interfaces.js'
 import type { WicImageCompareOptions } from './images.interfaces.js'
 import type { BoundingBoxes, ReportFileSizes } from './rectangles.interfaces.js'
 import type { FilePaths, FolderPaths } from 'src/base.interfaces.js'
