+++ packages/image-comparison-core/src/methods/createCompareReport.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { writeFileSync, readFileSync } from 'node:fs'
 import { createCompareReport, createJsonReportIfNeeded } from './createCompareReport.js'
-import type { CompareData } from '../resemble/compare.interfaces.js'
+import type { CompareData } from '../pixelmatch/compare.interfaces.js'
 import type { BoundingBox, IgnoreBoxes } from './rectangles.interfaces.js'
 import type { BaseDimensions } from '../base.interfaces.js'
 import { getBase64ScreenshotSize } from '../helpers/utils.js'
