+++ packages/image-comparison-core/tsconfig.json
@@ -2,10 +2,9 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
-    "allowJs": true,
     "rootDir": "./src",
     "outDir": "./dist",
     "baseUrl": "."
   },
-  "include": ["./src/**/*.ts", "./src/resemble/resemble.jimp.cjs"]
+  "include": ["./src/**/*.ts"]
 }
